# lab2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Приют и передача животных в семью",
    "description": "REST API приюта: животные, вольеры, заявки на усыновление, акты передачи, пользователи. Даты и время проставляет сервер, в запросах не передаются.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/users": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Список пользователей",
        "operationId": "listUsers",
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Зарегистрировать пользователя",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "201": {
            "description": "Ресурс создан",
            "headers": {
              "Location": {
                "description": "Адрес созданного ресурса",
                "style": "simple"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/enclosures": {
      "get": {
        "tags": [
          "Вольеры"
        ],
        "summary": "Список вольеров",
        "operationId": "listEnclosures",
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnclosureResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Вольеры"
        ],
        "summary": "Добавить вольер",
        "operationId": "createEnclosure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnclosureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "201": {
            "description": "Ресурс создан",
            "headers": {
              "Location": {
                "description": "Адрес созданного ресурса",
                "style": "simple"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EnclosureResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{id}": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Пользователь по идентификатору",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Удалить пользователя",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Частично обновить пользователя",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/enclosures/{id}": {
      "get": {
        "tags": [
          "Вольеры"
        ],
        "summary": "Вольер по идентификатору",
        "operationId": "getEnclosure",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EnclosureResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Вольеры"
        ],
        "summary": "Удалить вольер",
        "operationId": "deleteEnclosure",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "tags": [
          "Вольеры"
        ],
        "summary": "Частично обновить вольер",
        "operationId": "updateEnclosure",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnclosurePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Тело запроса не соответствует формату ресурса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Операция нарушает состояние данных",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EnclosureResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Тело ответа с ошибкой",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Момент формирования ответа",
            "example": "2026-09-26T07:07:27.667885Z"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "Код ответа",
            "example": 409
          },
          "error": {
            "type": "string",
            "description": "Причина, соответствующая коду",
            "example": "Conflict"
          },
          "message": {
            "type": "string",
            "description": "Описание причины ошибки"
          },
          "path": {
            "type": "string",
            "description": "Путь запроса",
            "example": "/api/animals/1"
          }
        }
      },
      "UserRequest": {
        "type": "object",
        "description": "Новый пользователь системы",
        "properties": {
          "username": {
            "type": "string",
            "description": "Логин",
            "example": "ivan",
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.-]{3,32}$"
          },
          "fullName": {
            "type": "string",
            "description": "Имя и фамилия",
            "example": "Иван Петров",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Адрес электронной почты",
            "example": "ivan@example.com",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "fullName",
          "username"
        ]
      },
      "UserResponse": {
        "type": "object",
        "description": "Пользователь системы",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Идентификатор",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "Логин",
            "example": "ivan"
          },
          "fullName": {
            "type": "string",
            "description": "Имя и фамилия",
            "example": "Иван Петров"
          },
          "email": {
            "type": "string",
            "description": "Адрес электронной почты",
            "example": "ivan@example.com"
          }
        }
      },
      "EnclosureRequest": {
        "type": "object",
        "description": "Новый вольер",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название вольера",
            "example": "Вольер для собак",
            "minLength": 1
          },
          "allowedSpecies": {
            "type": "array",
            "description": "Виды животных, которых допустимо размещать",
            "example": [
              "Собака"
            ],
            "items": {
              "type": "string",
              "minLength": 1
            },
            "minItems": 1,
            "uniqueItems": true
          },
          "capacity": {
            "type": "integer",
            "format": "int32",
            "description": "Вместимость вольера",
            "example": 4
          }
        },
        "required": [
          "allowedSpecies",
          "capacity",
          "name"
        ]
      },
      "EnclosureResponse": {
        "type": "object",
        "description": "Вольер приюта",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Идентификатор",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Название вольера",
            "example": "Вольер для собак"
          },
          "allowedSpecies": {
            "type": "array",
            "description": "Виды животных, которых допустимо размещать",
            "example": [
              "Собака"
            ],
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "capacity": {
            "type": "integer",
            "format": "int32",
            "description": "Вместимость вольера",
            "example": 4
          }
        }
      },
...
      "UserPatch": {
        "type": "object",
        "description": "Поля пользователя для изменения, отсутствующие поля не меняются",
        "properties": {
          "username": {
            "type": "string",
            "description": "Логин",
            "example": "ivan",
            "pattern": "^[a-zA-Z0-9_.-]{3,32}$"
          },
          "fullName": {
            "type": "string",
            "description": "Имя и фамилия",
            "example": "Иван Петров",
            "pattern": "\\S+"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Адрес электронной почты",
            "example": "ivan.petrov@example.com"
          }
        }
      },
      "EnclosurePatch": {
        "type": "object",
        "description": "Поля вольера для изменения, отсутствующие поля не меняются",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название вольера",
            "example": "Вольер для собак",
            "pattern": "\\S+"
          },
          "allowedSpecies": {
            "type": "array",
            "description": "Виды животных, которых допустимо размещать",
            "example": [
              "Собака",
              "Кошка"
            ],
            "items": {
              "type": "string",
              "pattern": "\\S+"
            },
            "uniqueItems": true
          },
          "capacity": {
            "type": "integer",
            "format": "int32",
            "description": "Вместимость вольера",
            "example": 6
          }
        }
      },
...
    }
  }
}